refactor: HangulUtil을 확장함수로 리팩터링
