+++ src/main/kotlin/dnd11th/blooming/api/service/guide/provider/ToxicityDetailProvider.kt
@@ -1,7 +1,7 @@
 package dnd11th.blooming.api.service.guide.provider
 
 import dnd11th.blooming.api.dto.guide.DetailToxicityResponse
-import dnd11th.blooming.common.util.HangulUtil.Companion.getEunOrNun
+import dnd11th.blooming.common.util.getEunOrNun
 import dnd11th.blooming.domain.entity.plant.Plant
 import dnd11th.blooming.domain.entity.plant.Toxicity
 import org.springframework.stereotype.Component
@@ -20,7 +20,7 @@ class ToxicityDetailProvider {
     }
 
     private fun makeDetailToxicityDescription(plant: Plant): String {
-        val eunOrNun = getEunOrNun(plant.korName)
+        val eunOrNun = plant.korName.getEunOrNun()
 
         val description =
             when (plant.toxicity) {
